paper-trader-league/services/data_ingest/requirements.txt
ccxt
pandas
sqlalchemy
psycopg2-binary
requests
feedparser
PyJWT[crypto]
